February 2022

+++ app/controllers/api/v1/lists/accounts_controller.rb
@@ -15,33 +15,59 @@ class Api::V1::Lists::AccountsController < Api::BaseController
   end
 
   def create
-    if Block.where(account_id: params[:account_id], target_account: current_account).exists?
-      raise GabSocial::NotPermittedError, 'Cannot add to list. That account has you blocked.'
-    else
-      ApplicationRecord.transaction do
-        @list.accounts << list_account
-      end
-      render_empty_success
+     # Must be list owner
+     if @list.account_id != current_account.id
+      raise GabSocial::NotPermittedError, 'Unable to perform action.'
     end
+
+    # Don't allow if blocked
+    if Block.where(account_id: params[:account_id], target_account: current_account).exists?
+      raise GabSocial::NotPermittedError, 'Cannot add to feed. That account has you blocked.'
+    end
+
+    # Don't allow if user is locked/private and current_account not following
+    if list_account.locked? && !current_account.following?(list_account)
+      raise GabSocial::NotPermittedError, 'Cannot add to feed. That account is private and you are not following them.'
+    end
+
+    # Don't allow if user already removed themself from this list
+    if ListRemovedAccount.where(list: @list, account_id: params[:account_id]).exists?
+      raise GabSocial::NotPermittedError, 'Cannot add to feed. That account has removed themself.'
+    end
+
+    ApplicationRecord.transaction do
+      @list.accounts << list_account
+    end
+
+    render_empty_success
   end
 
   def destroy
-    ListAccount.where(list: @list, account_id: params[:account_id]).destroy_all
+    # authorize either list owner of current account = params[:account_id]
+    if @list.account_id === current_account.id || params[:account_id] === current_account.id
+      ListAccount.where(list: @list, account_id: params[:account_id]).destroy_all
+      render_empty_success
+    else
+      raise GabSocial::NotPermittedError, 'Unable to perform action.'
+    end
+  end
+
+  def leave
+    ApplicationRecord.transaction do
+      @list.removed_accounts << current_account
+      ListAccount.where(list: @list, account_id: current_account.id).destroy_all
+    end
     render_empty_success
   end
 
   private
 
   def set_list
-    @list = List.where(account: current_account).find(params[:list_id])
+    @list = List.where(account: current_account).or(List.public_only).find(params[:list_id])
   end
 
   def load_accounts
-    if unlimited?
-      @list.accounts.includes(:account_stat).all
-    else
-      @list.accounts.includes(:account_stat).paginate_by_max_id(limit_param(DEFAULT_ACCOUNTS_LIMIT), params[:max_id], params[:since_id])
-    end
+    @list.accounts.includes(:account_stat).paginate_by_max_id(limit_param(DEFAULT_ACCOUNTS_LIMIT), params[:max_id], params[:since_id])
   end
 
   def list_account
@@ -57,16 +83,12 @@ class Api::V1::Lists::AccountsController < Api::BaseController
   end
 
   def next_path
-    return if unlimited?
-
     if records_continue?
       api_v1_list_accounts_url pagination_params(max_id: pagination_max_id)
     end
   end
 
   def prev_path
-    return if unlimited?
-
     unless @accounts.empty?
       api_v1_list_accounts_url pagination_params(since_id: pagination_since_id)
     end
@@ -88,7 +110,4 @@ class Api::V1::Lists::AccountsController < Api::BaseController
     params.slice(:limit).permit(:limit).merge(core_params)
   end
 
-  def unlimited?
-    params[:limit] == '0'
-  end
 end
